Rename EncryptSpringBootStarterTest
