+++ shardingsphere-features/shardingsphere-encrypt/shardingsphere-encrypt-spring/shardingsphere-encrypt-spring-boot-starter/src/test/resources/application-encrypt.properties
@@ -15,6 +15,9 @@
 # limitations under the License.
 #
 
-spring.shardingsphere.rules.encrypt.encryptors.aes_encryptor.type=test
-spring.shardingsphere.rules.encrypt.encryptors.aes_encryptor.props.aes.key.value=123456
+spring.shardingsphere.rules.encrypt.encryptors.aesEncryptor.type=AES
+spring.shardingsphere.rules.encrypt.encryptors.aesEncryptor.props.aes.key.value=123456
 
+spring.shardingsphere.rules.encrypt.tables.t_order.columns.pwd.cipher-column=pwd_cipher
+spring.shardingsphere.rules.encrypt.tables.t_order.columns.pwd.plain-column=pwd_plain
+spring.shardingsphere.rules.encrypt.tables.t_order.columns.pwd.encryptor-name=aesEncryptor
